+++ apis.tf
@@ -7,9 +7,12 @@ locals {
     "container.googleapis.com",
     "monitoring.googleapis.com",
     "logging.googleapis.com",
+    "servicenetworking.googleapis.com"
   ]
   _cloud_dns_api = var.google_services.use_cloud_dns ? ["dns.googleapis.com"] : []
-  enable_apis    = concat(local._required_apis, local._cloud_dns_api)
+  _cloud_sql_api = var.google_services.use_cloud_sql ? ["sqladmin.googleapis.com"] : []
+  enable_apis = concat(local._required_apis, local._cloud_dns_api,
+  local._cloud_sql_api)
 }
 
 resource "google_project_service" "openwisp_apis" {